- fix: offerbook filter (#242); Refactor: Moved Bottomsheet inside molecules/bottom_sheet
