+++ shared/presentation/src/commonMain/kotlin/network/bisq/mobile/presentation/ui/uicases/create_offer/CreateOfferPriceScreen.kt
@@ -117,7 +117,6 @@ fun CreateOfferTradePriceSelectorScreen() {
                 }
             }
 
-            // TODO:
             val tempStatus = "bisqEasy.price.feedback.sentence.some".i18n()
             NoteText(
                 notes = "bisqEasy.price.feedback.sentence".i18n(tempStatus),
